Moved `SettingsPanelObjectGrabber`
