+++ UITweaks/Installers/TweaksMenuInstaller.cs
@@ -1,5 +1,4 @@
 ﻿using UITweaks.UI;
-using UITweaks.Utilities;
 using UnityEngine;
 using Zenject;
